minor tmap_animation update suggestion
Hey Martijn,
I am just playing around with `tmap_animation()`. There is a minor probably Windows-only issue related to: 
https://stackoverflow.com/questions/40096966/r-tmap-animation-tmap-imagemagick
When installing ImageMagick, one has to check a box "install legacy files (e.g., convert.exe)". Not doing so results in an error message since there seems to be another convert command under Windows. This this is not well-documented, I think it would be more user-friendly to be more specific in the code of `tmap_animation()` in the sense to make clear that the convert command of ImageMagick should be used (see minor file change).
As a side note, the first example of `tmap_animation()` is not working, since there is no object called `NLD_prov`.